+++ lib/swid/frontend/dart/narrowStaticConstSyntacticEntity.dart
