+++ lib/swid/ir/util/extractStaticConstFromSyntacticEntity.dart
@@ -11,7 +11,7 @@ import 'package:hydro_sdk/swid/ir/swidStaticConstFunctionInvocation.dart';
 import 'package:hydro_sdk/swid/ir/swidStaticConstPrefixedExpression.dart';
 import 'package:hydro_sdk/swid/ir/swidStaticConstPrefixedIdentifier.dart';
 import 'package:hydro_sdk/swid/ir/swidStringLiteral.dart';
-import 'package:hydro_sdk/swid/ir/util/narrowStaticConstSyntacticEntity.dart';
+import 'package:hydro_sdk/swid/frontend/dart/narrowStaticConstSyntacticEntity.dart';
 
 SwidStaticConst extractStaticConstFromSyntacticEntity({@required SyntacticEntity syntacticEntity}) =>
     narrowStaticConstSyntacticEntity(
